+++ modules/@apostrophecms/area/ui/apos/components/AposAreaExpandedMenu.vue
@@ -103,6 +103,11 @@ export default {
         const group = this.createGroup(this.options.groups[item]);
         this.groups.push(group);
       }
+
+      const clipboard = this.getClipboard();
+      this.groups = clipboard
+        ? [ clipboard ].concat(this.groups)
+        : this.groups;
     } else if (this.options.widgets) {
       if (!this.isValidColumn(this.options.columns)) {
         console.warn(
@@ -121,6 +126,36 @@ export default {
   methods: {
     isValidColumn(count) {
       return count ? +count > 1 && +count < 4 : true;
+    },
+    getClipboard() {
+      const clipboard = apos.area.widgetClipboard.get();
+      if (!clipboard) {
+        return null;
+      }
+
+      const widgets = this.groups.flatMap(group => Object.values(group.widgets));
+      const matchingChoice = widgets.find(widget => widget.name === clipboard.type);
+      if (!matchingChoice) {
+        return null;
+      }
+
+      const group = {
+        label: 'Clipboard',
+        widgets: [
+          {
+            type: 'clipboard',
+            ...matchingChoice,
+            label: {
+              key: 'apostrophe:pasteWidget',
+              widget: this.$t(matchingChoice.label)
+            },
+            clipboard
+          }
+        ],
+        columns: 1
+      };
+
+      return group;
     },
     createGroup(config) {
       const group = {